+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/NewShop.kt
@@ -24,6 +24,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.example.compose_study.ui.theme.ComposeStudyTheme
+import com.example.compose_study.utils.ContentsDivider
 
 @Composable
 fun NewShopScreen() {
