+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/ReReservation.kt
@@ -32,6 +32,9 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.example.compose_study.R
 import com.example.compose_study.ui.theme.ComposeStudyTheme
+import com.example.compose_study.utils.ui.Category
+import com.example.compose_study.utils.ContentsDivider
+import com.example.compose_study.utils.ui.LineIndicatorTabs
 
 @Composable
 fun ReReservationScreen(
@@ -67,7 +70,7 @@ fun ReReservationScreen(
         modifier = Modifier.fillMaxWidth().background(Color.White)
     ) {
         ReReservationTitle()
-        BottomIndicatorTabRow(
+        LineIndicatorTabs(
             tabs = tabs,
             selectedTabIndex = selectedTab,
             onTabClicked = { tabIndex, _ -> onTabClicked(tabIndex) }
