+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/RecommendMenu.kt
@@ -31,6 +31,9 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.example.compose_study.R
 import com.example.compose_study.ui.theme.ComposeStudyTheme
+import com.example.compose_study.utils.ui.Category
+import com.example.compose_study.utils.ContentsDivider
+import com.example.compose_study.utils.ui.LineIndicatorTabs
 
 @Composable
 fun RecommendMenuScreen(
@@ -81,7 +84,7 @@ fun RecommendMenuScreen(
         modifier = Modifier.fillMaxWidth().background(Color.White).clickable{ onMenuClicked() }
     ) {
         RecommendMenuTitle()
-        BottomIndicatorTabRow(
+        LineIndicatorTabs(
             tabs = tabs,
             selectedTabIndex = selectIndex,
             onTabClicked = { tabIndex, _ -> selectIndex = tabIndex }
