+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/RecommendNail.kt
@@ -29,6 +29,7 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.example.compose_study.R
 import com.example.compose_study.ui.theme.ComposeStudyTheme
+import com.example.compose_study.utils.ContentsDivider
 
 @Composable
 fun RecommendNailScreen() {
