+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/RecommendStyle.kt
@@ -33,6 +33,9 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.example.compose_study.R
 import com.example.compose_study.ui.theme.ComposeStudyTheme
+import com.example.compose_study.utils.ui.Category
+import com.example.compose_study.utils.ContentsDivider
+import com.example.compose_study.utils.ui.LineIndicatorTabs
 
 @Composable
 fun RecommendStyleScreen() {
@@ -68,7 +71,7 @@ fun RecommendStyleScreen() {
             .background(Color.White)
     ) {
         RecommendStyleTitle()
-        BottomIndicatorTabRow(
+        LineIndicatorTabs(
             tabs = tabs,
             selectedTabIndex = selectIndex,
             onTabClicked = { tabIndex, _ -> selectIndex = tabIndex }
