+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/UpdateProfile.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.compose_study.R
+import com.example.compose_study.utils.ContentsDivider
 
 @Composable
 fun UpdateProfileCard() {
