+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/loading/LoadingBannerScreen.kt
@@ -9,9 +9,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.example.compose_study.utils.LoadingCornerRounded
-import com.example.compose_study.utils.LoadingFillMaxWidth
-import com.example.compose_study.utils.VerticalDivider
+import com.example.compose_study.utils.ui.LoadingCornerRounded
+import com.example.compose_study.utils.ui.LoadingFillMaxWidth
+import com.example.compose_study.utils.ui.VerticalDivider
 
 @Composable
 fun LoadingTopBannerScreen(
