+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/loading/LoadingLocationScreen.kt
@@ -8,11 +8,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.compose_study.utils.ColorLineDivider
-import com.example.compose_study.utils.HorizontalDivider
-import com.example.compose_study.utils.LoadingCircleShape
-import com.example.compose_study.utils.LoadingCornerRounded
-import com.example.compose_study.utils.VerticalDivider
+import com.example.compose_study.utils.ui.ColorLineDivider
+import com.example.compose_study.utils.ui.HorizontalDivider
+import com.example.compose_study.utils.ui.LoadingCircleShape
+import com.example.compose_study.utils.ui.LoadingCornerRounded
+import com.example.compose_study.utils.ui.VerticalDivider
 
 @Composable
 fun LoadingLocationScreen(
