+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/loading/LoadingMenuScreen.kt
@@ -8,9 +8,9 @@ import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.compose_study.utils.HorizontalDivider
-import com.example.compose_study.utils.LoadingCornerRounded
-import com.example.compose_study.utils.VerticalDivider
+import com.example.compose_study.utils.ui.HorizontalDivider
+import com.example.compose_study.utils.ui.LoadingCornerRounded
+import com.example.compose_study.utils.ui.VerticalDivider
 
 @Composable
 fun LoadingMenuScreen(
