+++ app/src/main/java/com/example/compose_study/ui/screen/feature/component/loading/LoadingScreen.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.example.compose_study.utils.VerticalDivider
+import com.example.compose_study.utils.ui.VerticalDivider
 
 @Composable
 fun LoadingScreen(
