+++ app/src/main/java/com/example/compose_study/ui/screen/stylebook/StyleBookHeader.kt
@@ -20,7 +20,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.compose_study.R
-import com.example.compose_study.utils.ColorLineDivider
 
 @Composable
 fun StyleBookHeader(
